+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
+* Add conditional contents.
+
 ## 1.2.0 (May. 30, 2016)
 
 * Allow programmatic use.
-* Tolerate missing files (don't crash!)
-* Write missing file refs to output (disabled by default), with the `--missing-refs` option
+* Tolerate missing files (don't crash!).
+* Write missing file refs to output (disabled by default), with the `--missing-refs` option.
 
 ## 1.1.0 (May. 20, 2016)