+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 * Add conditional contents.
 
-## 1.2.0 (May. 30, 2016)
+## 1.4.0 (Jan. 20, 2017)
+
+* Remove unnecessary dependencies.
+
+## 1.2.0, 1.3.0 (May. 30, 2016)
 
 * Allow programmatic use.
 * Tolerate missing files (don't crash!).